I can't do it :(
I will always need a full scan, which is at least O(n): changes would say added/removed values, but not reordered ones. And when a new value is added, I need to go from minIdx to the end to shift every value... which is O(n)

In sort I don't have this problem because it doesn't matter if the input gets reordered: only on add / change / remove, and that has a cost O(1) or O(logn). It's still technically O(n) because reloading indices